+++ frontend/ogy_dashboard/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### [1.0.7] - 2024-06-12
 
+#### Fix
+
+- Inacurate OGY and ICP treasury account data
+
 #### Added
 
 - Proposals metrics on Governance page.
@@ -18,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Improve display footer notice.
 - Transactions table accounts columns display.
+- Inacurate OGY treasury account
 
 #### Added